+++ src/main/scala/pl/edu/pw/elka/adoptimizer/categorization/classifier/TextClassifier.scala
@@ -7,7 +7,7 @@ import pl.edu.pw.elka.adoptimizer.categorization.model.Sample
  */
 
 trait TextClassifier {
-  def classify(sample: Sample): Double
+  def classify(sample: Sample): Map[String, Double]
   def fit(samples: List[Sample]): Unit
   def save(): Any
   def load(state: Any): Unit
