+++ src/main/scala/pl/edu/pw/elka/adoptimizer/categorization/tokenizer/Tokenizer.scala
@@ -2,9 +2,6 @@ package pl.edu.pw.elka.adoptimizer.categorization.tokenizer
 
 import pl.edu.pw.elka.adoptimizer.categorization.model.Document
 
-/**
- * Created by leszek on 26/12/2017.
- */
 trait Tokenizer {
   def tokenize(text: String): Document
 }
